requirements.txt
streamlit
bcrypt
psycopg2-binary
streamlit-cookies-manager
numpy  # Optional, if needed
pandas  # Optional, if needed
